wheels-wins-data-collector/requirements 2.txt
# Standalone Data Collection System Requirements
# Safe to delete after use

# Web scraping and HTTP
aiohttp>=3.8.0
beautifulsoup4>=4.12.0
requests>=2.31.0
httpx>=0.24.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0
geopandas>=0.13.0
geopy>=2.3.0

# Database integration
supabase>=2.0.0
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0

# Configuration and logging
pyyaml>=6.0
python-dotenv>=1.0.0
loguru>=0.7.0

# Data validation and enhancement
pydantic>=2.0.0
openai>=1.0.0
googlemaps>=4.10.0

# Utility libraries
tqdm>=4.65.0
click>=8.1.0
python-dateutil>=2.8.0
pytz>=2023.3

# API integrations
overpy>=0.6  # OpenStreetMap
osmapi>=3.1.0  # OSM API

# Rate limiting and retries
tenacity>=8.2.3
ratelimit>=2.2.1

# Progress tracking
redis>=5.0.1  # For distributed progress tracking
schedule>=1.2.0  # For internal scheduling

# Monitoring
sentry-sdk>=1.38.0  # Error tracking

# Optional: For advanced data processing
scikit-learn>=1.3.0  # For clustering and deduplication
nltk>=3.8.0          # For text processing
spacy>=3.6.0         # For location extraction